requirements.txt
fpdf2
imageio
requests